Non-destructive topology creation.
This fixes #2, changing topojson.topology so that it no longer destroys the
input objects. It also fixes #3 by removing type reduction: for example, an
empty MultiLineString geometry on input is now an empty MultiLineString geometry
on output, rather than a null geometry.

Most of the internal implementation is still destructive. This works by changing
the geometry method to make a shallow copy of the input, and by changing the
prequantize method to make a copy of the coordinates.

I’d like to improve this so that prequantize operates on the extracted
coordinates, so that we can avoid the additional copy. This should be possible
in a future commit. (It would also be nice if the intermediate representation
used two typed arrays rather than an array of objects, but I think that will be
more difficult to implement.)